Added Jest coverage configuration
If the goal of enabling coverage is to see the parts of the codebase that lack testing then we need to configure jest to do this. Adding the --coverage flag in the test script will display the coverage table, however it will only show files for which tests have been written. I added in a jest config script that monitors any js file in the src directory or it's sub-directories.